Renamed WLAN Module into WIFI
