+++ lib/WIFI/WIFI.cpp
@@ -26,7 +26,7 @@
 * @author: Gabryel Reyes <gabryelrdiaz@gmail.com>
 **************************************************************************************************/
 /* INCLUDES **************************************************************************************/
-#include "WLAN.h"
+#include "WIFI.h"
 #include <EEPROM.h>
 
 /* CONSTANTS *************************************************************************************/
