+++ lib/WebServer/WebServer.h
@@ -29,7 +29,7 @@
 #define WEB_SERVER_H_
 
 /* INCLUDES **************************************************************************************/
-#include <WLAN.h>
+#include "WIFI.h"
 #include <ESP8266WebServer.h>
 #include <ESP8266mDNS.h>
 #include <DNSServer.h>
